securevault/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
requests==2.31.0
Werkzeug==2.3.7
python-dotenv==1.0.0
